Authorization bug fixed

+++ server/controllers/user.js
@@ -295,6 +295,17 @@ exports.postComment = async (req, res) => {
 };
 
 exports.deleteComment = async (req, res) => {
+  const comment = await Comment.findById(req.params.id);
+
+  if (!comment) {
+    return res.status(404).json('Comment not found');
+  }
+
+  // Check if the user is the author of the comment
+  if (!comment.user !== req.user._id) {
+    return res.status(403).json('You are not authorized to delete this comment');
+  }
+
   await Comment.findByIdAndDelete(req.params.id);
   return res.status(200).json('Comment Deleted');
 };
